hmis-backend/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hmis-backend"
version = "1.0.0"
description = "HMIS SaaS - Sistema de Gestion Hospitalaria Cloud-Native para Latinoamerica"
requires-python = ">=3.12"
license = {text = "Proprietary"}
authors = [
    {name = "HMIS Team"},
]
dependencies = [
    # Framework Web
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6",
    # Base de Datos
    "sqlalchemy[asyncio]>=2.0.25",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    # Validacion y Serializacion
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",
    "email-validator>=2.1.0",
    # Autenticacion y Seguridad
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "argon2-cffi>=23.1.0",
    # Cache y Colas
    "redis>=5.0.0",
    # HTTP Client
    "httpx>=0.26.0",
    # Utilidades
    "python-dateutil>=2.8.2",
    "pytz>=2024.1",
    "babel>=2.14.0",
    # Busqueda
    "meilisearch>=0.31.0",
    # Almacenamiento
    "boto3>=1.34.0",
    # Monitoreo
    "sentry-sdk[fastapi]>=1.40.0",
    "prometheus-client>=0.20.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.26.0",
    "factory-boy>=3.3.0",
    "faker>=22.0.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
]

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "C4", "SIM", "TCH"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
filterwarnings = ["ignore::DeprecationWarning"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
